Add files via upload
This machine learning project focuses on predicting the type of traffic violation committed by drivers. The project aims to utilize several machine learning models to identify the most accurate one to make predictions. The dataset contains various independent variables, including the time of the stop, driver's gender, age, race, whether a search was conducted, the outcome of the stop, whether the driver was arrested or not, the duration of the stop, drug-related stops, year, and month. The target variable is the type of traffic violation committed. By using the collected data, the project will develop a model that can accurately predict the type of traffic violation committed, helping to inform law enforcement and promote safe driving practices.


The business problem in this machine learning project could be improving traffic safety and reducing the number of traffic violations. By accurately predicting the type of traffic violation committed, law enforcement agencies can better target their resources and efforts to deter and prevent specific types of violations. This can ultimately lead to a safer driving environment and help reduce the number of accidents and injuries caused by traffic violations. Additionally, the project could help identify trends in traffic violations, which can inform policy decisions and interventions aimed at reducing specific types of violations. Therefore, the business problem that this project aims to solve is improving traffic safety and reducing the number of traffic violations.